Move cards render logic to the trip-controller
